requirements.txt
anthropic>=0.39.0
solana>=0.30.0
aiohttp>=3.9.0
pyyaml>=6.0
python-dotenv>=1.0.0
websockets>=12.0
